refactor(auth,verification): move email verification service to `auth` package
