+++ backend/internal/verification/handler.go
@@ -2,7 +2,6 @@ package verification
 
 import (
 	"errors"
-	"log/slog"
 	"net/http"
 
 	"github.com/vexgo-org/vexgo/backend/internal/middleware"
@@ -19,43 +18,6 @@ type Handler struct {
 // NewHandler creates a verification HTTP handler with the given dependencies.
 func NewHandler(deps Deps) *Handler {
 	return &Handler{svc: NewService(deps), mw: middleware.NewAuth(deps.DB, deps.JWTSecret)}
-}
-
-// VerifyEmail verifies email (supports initial verification and email change)
-func (h *Handler) VerifyEmail(c *gin.Context) {
-	token := c.Query("token")
-	slog.Debug("email verification request received", "hasToken", token != "")
-
-	if token == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Verification token cannot be empty"})
-		return
-	}
-
-	emailChange, newEmail, err := h.svc.VerifyEmail(c.Request.Context(), token)
-	if err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-
-	if emailChange {
-		if newEmail != "" {
-			c.JSON(http.StatusOK, gin.H{
-				"message":         "Email change successful! Your new email is now active.",
-				"require_relogin": true,
-				"new_email":       newEmail,
-			})
-		} else {
-			c.JSON(http.StatusOK, gin.H{
-				"message":         "Email change successful! Your new email is now active.",
-				"require_relogin": true,
-			})
-		}
-		return
-	}
-
-	c.JSON(http.StatusOK, gin.H{
-		"message": "Email verification successful! You can now log in.",
-	})
 }
 
 // GetVerificationStatus gets current user's email verification status
