+++ backend/internal/verification/routes.go
@@ -8,8 +8,6 @@ import (
 // Route paths and middleware chains are identical to the original registration
 // in the legacy handler package.
 func (h *Handler) RegisterRoutes(api *gin.RouterGroup) {
-	api.GET("/verify-email", h.VerifyEmail)
-
 	api.GET("/captcha", h.GenerateCaptcha)
 	api.POST("/captcha/verify", h.VerifyCaptcha)
